Refactory on k8s unit tests (#90)
* Refactory on k8s unit tests


* Fixed: Refactory on k8s unit tests


* Fixed: Refactory on k8s unit tests for windows


* Fixed: Refactory on k8s unit tests for windows 0


* Fixed: Refactory on k8s unit tests for windows 1
